multiple flock enhancement
 Flock Class: The Flock class is designed to hold properties of a flock, such as alignment weight, cohesion weight, separation weight, initial energy, and particle count.

Flocks List: The GameManager class has a List<Flock> flocks which is meant to store multiple instances of the Flock class, allowing for multiple flocks to be managed.

AddFlock Method: The AddFlock method in the GameManager class allows for the creation of a new flock and adds it to the flocks list.

SwitchToFlock Method: The SwitchToFlock method allows you to switch the current active flock to any flock in the flocks list based on an index.

SaveCurrentPopulation & LoadPopulation Methods: These methods in the GameManager class allow for saving the properties of the current flock and loading properties of a selected flock, respectively.

Population Management in InputManager: The InputManager class has methods like LoadSelectedPopulation and UpdatePopulationList which are designed to interact with the multiple flocks in the flocks list.